// src/enums/Tag.enum.ts
export enum Tag {
  AIRBYTE = "AIRBYTE",
  AIRTABLE = "AIRTABLE",
  ANGULAR = "ANGULAR",
  BOOTSTRAP = "BOOTSTRAP",
  C = "C",
  CPP = "C++",
  CSS = "CSS",
  DAGSTER = "DASGTER",
  DBT = "DBT",
  EXPRESS = "EXPRESS",
  GIT = "GIT",
  HTML = "HTML",
  JAVA = "JAVA",
  JAVASCRIPT = "JAVASCRIPT",
  JEST = "JEST",
  KERAS = "KERAS",
  MATLAB = "MATLAB",
  METABASE = "METABASE",
  MONGO = "MONGO",
  MONGOOSE = "MONGOOSE",
  NODE = "NODE",
  POSTGRES = "POSTGRES",
  PYTHON = "PYTHON",
  SUPERSET = "SUPERSET",
  TYPESCRIPT = "TYPESCRIPT",
}